把副程式Find 的輸入數字宣告改為 unsigned long long

+++ DSlex1_31_10720102_10720127.cpp
@@ -20,7 +20,7 @@ using namespace std;
 
 
 // 此為的一找出數列中的最小因數差之函數
-void Find( int target ){
+void Find( unsigned long long target ){
 
 
 // 設一數字num 為target 的開根號數
